+++ src/init/utils.ts
@@ -12,12 +12,13 @@ import { putLogEvent } from '../logger';
 import { generateSettings } from "../settings";
 import { pathUtils } from '../utils/path_utils';
 import { checkServerDown } from '../utils/api_utils';
-import { IFileToUpload, IUserPlan } from '../interface';
+import { IFileToUpload } from '../interface';
 import { trackRepoHandler } from '../handlers/commands_handler';
 import { uploadFileTos3, uploadRepoToServer } from '../utils/upload_utils';
-import { CONNECTION_ERROR_MESSAGE, DIFF_SOURCE, NOTIFICATION } from '../constants';
+import { CONNECTION_ERROR_MESSAGE, DIFF_SOURCE, NOTIFICATION, RETRY_BRANCH_SYNC_AFTER } from '../constants';
 import { getSkipRepos, isRepoActive, readYML, getSyncIgnoreItems } from '../utils/common';
 import { getPlanLimitReached } from '../utils/pricing_utils';
+import { CodeSyncState, CODESYNC_STATES } from '../utils/state_utils';
 
 export class initUtils {
 	repoPath: string;
@@ -30,21 +31,10 @@ export class initUtils {
 		this.settings = generateSettings();
 	}
 
-	isValidRepoSize (syncSize: number, userPlan: IUserPlan)  {
-		const isValid = userPlan.SIZE >= syncSize;
-		if (!isValid) {
-			vscode.window.showErrorMessage(`${NOTIFICATION.REPOS_LIMIT_BREACHED} ${userPlan.SIZE}`);
-		}
-		return isValid;
-	}
-
-	isValidFilesCount (filesCount: number, userPlan: IUserPlan) {
-		const isValid = userPlan.FILE_COUNT >= filesCount;
-		if (!isValid) {
-			vscode.window.showErrorMessage(`${NOTIFICATION.FILES_LIMIT_BREACHED}\n
-			You can add only ${userPlan.FILE_COUNT} files (Trying to add ${filesCount} files)`);
-		}
-		return isValid;
+	isBranchSyncInProcess (syncingBranchKey: string) {
+		const branchSyncStartedAt = CodeSyncState.get(syncingBranchKey);
+		const isSyncInProcess = branchSyncStartedAt && (new Date().getTime() - branchSyncStartedAt) < RETRY_BRANCH_SYNC_AFTER;
+		return isSyncInProcess;
 	}
 
 	isSyncAble(relPath: string) {
@@ -53,14 +43,12 @@ export class initUtils {
 		return !ig.ignores(relPath);
 	}
 
-	getSyncablePaths (userPlan: IUserPlan, isPopulatingBuffer = false) {
+	getSyncablePaths () {
 		const itemPaths: IFileToUpload[] = [];
 		const repoPath = this.repoPath;
 		const viaDaemon = this.viaDaemon;
 		const syncIgnoreItems = getSyncIgnoreItems(repoPath);
 
-		let syncSize = 0;
-		let limitReached = false;
 		const skipRepos = getSkipRepos(repoPath, syncIgnoreItems);
 
 		const options = {
@@ -80,19 +68,13 @@ export class initUtils {
 							created_at: fileStats.ctime,
 							modified_at: fileStats.mtime
 						});
-						syncSize += fileStats.size;
-					}
-					if (!viaDaemon && !isPopulatingBuffer &&
-						(!self.isValidRepoSize(syncSize, userPlan) ||
-						!self.isValidFilesCount(itemPaths.length, userPlan))) {
-						limitReached = true;
 					}
 					next();
 				}
 			}
 		};
 		walk.walkSync(repoPath, options);
-		return limitReached ? [] : itemPaths;
+		return itemPaths;
 	}
 
 	copyFilesTo(filePaths: string[], destination: string, useFormattedRepoPath = false) {
@@ -176,7 +158,7 @@ export class initUtils {
 		fs.writeFileSync(this.settings.CONFIG_PATH, yaml.safeDump(configJSON));
 	}
 
-	async uploadRepoToS3(branch: string, token: string, uploadResponse: any) {
+	async uploadRepoToS3(branch: string, token: string, uploadResponse: any, syncingBranchKey: string) {
 		const viaDaemon = this.viaDaemon;
 		const s3Urls =  uploadResponse.urls;
 		const tasks: any[] = [];
@@ -200,14 +182,18 @@ export class initUtils {
 			function (err, results) {
 				// the results array will equal ['one','two'] even though
 				// the second function had a shorter timeout.
-				if (err) return;
+				if (err) {
+					CodeSyncState.set(syncingBranchKey, false);
+					return;
+				}
 				// delete .originals repo
 				if (fs.existsSync(originalsRepoBranchPath)) {
 					fs.rmdirSync(originalsRepoBranchPath, { recursive: true });
 				}
 				// Hide Connect Repo
 				vscode.commands.executeCommand('setContext', 'showConnectRepoView', false);
-
+				// Reset key for syncingBranch
+				CodeSyncState.set(syncingBranchKey, false);
 				// Show success notification
 				if (!viaDaemon) {
 					vscode.window.showInformationMessage(NOTIFICATION.REPO_SYNCED, ...[
@@ -224,10 +210,9 @@ export class initUtils {
 
 	async uploadRepo(branch: string, token: string, itemPaths: IFileToUpload[],
 					userEmail: string, isPublic=false) {
-
 		// Check plan limits
 		const { planLimitReached, canRetry } = getPlanLimitReached();
-		if (planLimitReached && !canRetry) return;
+		if (planLimitReached && !canRetry) return false;
 					
 		const repoName = path.basename(this.repoPath);
 		const configJSON = readYML(this.settings.CONFIG_PATH);
@@ -243,7 +228,7 @@ export class initUtils {
 				created_at: new Date(fileToUpload.created_at).getTime() / 1000
 			};
 		});
-
+		
 		if (!repoInConfig) {
 			configJSON.repos[this.repoPath] = {
 				branches: {},
@@ -259,8 +244,16 @@ export class initUtils {
 		const isServerDown = await checkServerDown();
 		if (isServerDown) {
 			if (!this.viaDaemon) putLogEvent(CONNECTION_ERROR_MESSAGE);
-			return;
+			return false;
 		}
+
+		// Check if branch is already being synced, skip it
+		const syncingBranchKey = `${CODESYNC_STATES.SYNCING_BRANCH}:${repoName}:${branch}`;
+		const isSyncInProcess = this.isBranchSyncInProcess(syncingBranchKey);
+		if (isSyncInProcess) return false;
+
+		// Set key here that Branch is being synced
+		CodeSyncState.set(syncingBranchKey, new Date().getTime());
 
 		console.log(`Uploading new branch: ${branch} for repo: ${this.repoPath}`);
 
@@ -273,16 +266,16 @@ export class initUtils {
 			platform: os.platform()
 		};
 
-		// Set key here that Branch is being synced
 		const json = await uploadRepoToServer(token, data);
 		if (json.error) {
 			// Reset the key here and try again in next attempt
+			CodeSyncState.set(syncingBranchKey, false);
 			const error = this.viaDaemon ? NOTIFICATION.ERROR_SYNCING_BRANCH : NOTIFICATION.ERROR_SYNCING_REPO;
 			putLogEvent(error, userEmail, json.error);
 			if (!this.viaDaemon) {
 				vscode.window.showErrorMessage(NOTIFICATION.SYNC_FAILED);
 			}
-			return;
+			return false;
 		}
 		/*
 			Response from server looks like
@@ -311,6 +304,8 @@ export class initUtils {
 		this.saveFileIds(branch, token, user.email, json.response);
 
 		// Upload to s3
-		await this.uploadRepoToS3(branch, token, json.response);
+		await this.uploadRepoToS3(branch, token, json.response, syncingBranchKey);
+
+		return true;
 	}
 }