+++ src/interface.ts
@@ -32,15 +32,7 @@ export interface IFileToUpload {
 }
 
 export interface IUser {
-	email: string,
-	plan: IUserPlan,
-	repo_count: number
-}
-
-export interface IUserPlan {
-	SIZE: number;
-	FILE_COUNT: number;
-	REPO_COUNT: number;
+	email: string
 }
 
 export interface IUserProfile {