+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.6.4] - 2022-08-30
+### Updated
+- Fixed multiple API calls for branch upload from daemon
+- Handling server side Plan Limit errors upon repo/branch sync
+
 ## [3.6.3] - 2022-08-13
 ### Updated
 - Using wss for secure websocket connection